requirements.txt
asgiref==3.8.1
Django==5.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
sqlparse==0.4.4
tzdata==2024.1
